# java/src/main/java/se/gewalli/polymorphic_swagger/api/OrderApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.2.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package se.gewalli.polymorphic_swagger.api;

import se.gewalli.polymorphic_swagger.model.AddOrder;
import se.gewalli.polymorphic_swagger.model.AddProductToOrder;
import java.util.Map;
import se.gewalli.polymorphic_swagger.model.OrderModel;
import se.gewalli.polymorphic_swagger.model.ProblemDetails;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import javax.validation.Valid;

import java.math.BigInteger;
import java.util.List;
import java.util.concurrent.CompletableFuture;

@Validated
@Tag(name = "order-api", description = "the Order API")
public interface OrderApi {

    /**
     * GET /api/v1/orders
     *
     * @return Success (status code 200)
     */
    @Operation(operationId = "apiV1OrdersGet", tags = { "Orders" }, responses = {
            @ApiResponse(responseCode = "200", description = "Success", content = {
                    @Content(mediaType = "text/plain", schema = @Schema(implementation = OrderModel.class)),
                    @Content(mediaType = "application/json", schema = @Schema(implementation = OrderModel.class)),
                    @Content(mediaType = "text/json", schema = @Schema(implementation = OrderModel.class))
            })
    })
    @RequestMapping(method = RequestMethod.GET, value = "/api/v1/orders", produces = { "text/plain", "application/json",
            "text/json" })
    CompletableFuture<ResponseEntity<List<OrderModel>>> getOrders(

    );

    /**
     * GET /api/v1/orders/{id}
     *
     * @param id (required)
     * @return Not Found (status code 404)
     *         or Success (status code 200)
     */
    @Operation(operationId = "apiV1OrdersIdGet", tags = { "Orders" }, responses = {
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                    @Content(mediaType = "text/plain", schema = @Schema(implementation = ProblemDetails.class)),
                    @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetails.class)),
                    @Content(mediaType = "text/json", schema = @Schema(implementation = ProblemDetails.class))
            }),
            @ApiResponse(responseCode = "200", description = "Success", content = {
                    @Content(mediaType = "text/plain", schema = @Schema(implementation = OrderModel.class)),
                    @Content(mediaType = "application/json", schema = @Schema(implementation = OrderModel.class)),
                    @Content(mediaType = "text/json", schema = @Schema(implementation = OrderModel.class))
            })
    })
    @RequestMapping(method = RequestMethod.GET, value = "/api/v1/orders/{id}", produces = { "text/plain",
            "application/json", "text/json" })
    CompletableFuture<ResponseEntity<OrderModel>> getOrder(
            @Parameter(name = "id", description = "", required = true) @PathVariable("id") BigInteger id);

    /**
     * POST /api/v1/orders/{id}/products :
     * 
     *
     * @param id                order id (required)
     * @param addProductToOrder (optional)
     * @return Bad Request (status code 400)
     */
    @Operation(operationId = "apiV1OrdersIdProductsPost", summary = "", tags = { "Orders" }, responses = {
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                    @Content(mediaType = "text/plain", schema = @Schema(implementation = Map.class)),
                    @Content(mediaType = "application/json", schema = @Schema(implementation = Map.class)),
                    @Content(mediaType = "text/json", schema = @Schema(implementation = Map.class))
            })
    })
    @RequestMapping(method = RequestMethod.POST, value = "/api/v1/orders/{id}/products", produces = { "text/plain",
            "application/json", "text/json" }, consumes = { "application/json", "text/json", "application/*+json" })
    CompletableFuture<ResponseEntity<OrderModel>> addProductToOrder(
            @Parameter(name = "id", description = "order id", required = true) @PathVariable("id") BigInteger id,
            @Parameter(name = "AddProductToOrder", description = "") @Valid @RequestBody(required = false) AddProductToOrder addProductToOrder);

    /**
     * POST /api/v1/orders :
     * 
     *
     * @param addOrder (optional)
     * @return Bad Request (status code 400)
     */
    @Operation(operationId = "apiV1OrdersPost", summary = "", tags = { "Orders" }, responses = {
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                    @Content(mediaType = "text/plain", schema = @Schema(implementation = Map.class)),
                    @Content(mediaType = "application/json", schema = @Schema(implementation = Map.class)),
                    @Content(mediaType = "text/json", schema = @Schema(implementation = Map.class))
            })
    })
    @RequestMapping(method = RequestMethod.POST, value = "/api/v1/orders", produces = { "text/plain",
            "application/json", "text/json" }, consumes = { "application/json", "text/json", "application/*+json" })
    CompletableFuture<ResponseEntity<OrderModel>> addOrder(
            @Parameter(name = "AddOrder", description = "") @Valid @RequestBody(required = false) AddOrder addOrder);

}